+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 1. Added `Client.delta_time_modifier` to manipulate the game's clockspeed.
 2. Added `Client.raw_delta_time` to fetch the client's delta time without modifiers.
 3. Added `Entity.static_rect` which is a `pygame.Rect` that won't change size due to rotations.
+4. Added `Entity.rotation_offset`.
 
 ### New Improvements
 1. Working `pygame.SCALED` replacement so that the `Client.screen` will no longer have black bars upon screen resize.